requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
backports.tarfile==1.2.0
backports.zoneinfo==0.2.1
blinker==1.8.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
Django==4.2
django-extensions==3.2.3
django-shell-plus==1.1.7
djangorestframework==3.15.2
Flask==2.3.3
frozenlist==1.4.1
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
Jinja2==3.1.4
keyring==25.2.1
MarkupSafe==2.1.5
more-itertools==10.3.0
multidict==6.0.5
pillow==10.4.0
psycopg2==2.9.9
psycopg2-binary==2.8.5
pycparser==2.22
pyee==11.1.0
PyJWT==1.7.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
yarl==1.9.4
zipp==3.19.2
